+++ apps/nextjs/tests-e2e/tests/teaching-materials/helpers.ts
@@ -1,5 +1,3 @@
-import { aiLogger } from "@oakai/logger";
-
 import type { Page } from "@playwright/test";
 
 import {
@@ -10,18 +8,10 @@ import {
   updateMaterialSessionResponse,
 } from "./fixtures";
 
-const log = aiLogger("teaching-materials:testing");
-
 export const applyTeachingMaterialsMockAPIRequests = async (page: Page) => {
   await page.route(
     "**/api/trpc/main/teachingMaterials.generatePartialLessonPlanObject?batch=1",
-    async (route, request) => {
-      const postData = request.postDataJSON();
-      log.info(
-        "Intercepted generatePartialLessonPlanObject tRPC call with params:",
-        postData,
-      );
-
+    async (route) => {
       await route.fulfill({
         status: 200,
         contentType: "application/json",
@@ -30,16 +20,9 @@ export const applyTeachingMaterialsMockAPIRequests = async (page: Page) => {
     },
   );
 
-  // Intercept createMaterialSession calls
   await page.route(
     "**/api/trpc/main/teachingMaterials.createMaterialSession?batch=1",
-    async (route, request) => {
-      const postData = request.postDataJSON();
-      log.info(
-        "Intercepted createMaterialSession tRPC call with params:",
-        postData,
-      );
-
+    async (route) => {
       await route.fulfill({
         status: 200,
         contentType: "application/json",
@@ -48,16 +31,9 @@ export const applyTeachingMaterialsMockAPIRequests = async (page: Page) => {
     },
   );
 
-  // Intercept updateMaterialSession calls
   await page.route(
     "**/api/trpc/main/teachingMaterials.updateMaterialSession?batch=1",
-    async (route, request) => {
-      const postData = request.postDataJSON();
-      log.info(
-        "Intercepted updateMaterialSession tRPC call with params:",
-        postData,
-      );
-
+    async (route) => {
       await route.fulfill({
         status: 200,
         contentType: "application/json",
@@ -71,10 +47,6 @@ export const applyTeachingMaterialsMockAPIRequests = async (page: Page) => {
     async (route, request) => {
       const postData: { json: { adaptsOutputId: string } }[] =
         request.postDataJSON() ?? [];
-      log.info(
-        "Intercepted generateTeachingMaterial tRPC call with params:",
-        postData,
-      );
 
       const body = postData[0]?.json?.adaptsOutputId
         ? generateTeachingMaterialWithAdaptsOutputId